Add Actor
A new type of entity was created. This entities have Fighter and AI
components that allow them to move around and fight. Nevertheless,
no damage dealing is introduced and hence fighting is not happening
for now. Actors were added to the engine to allow actors take turns and
were added to the map to distinguish them from other entities.

All based on: http://rogueliketutorials.com/tutorials/tcod/v2/part-6/