Add damage dealing, fix minor bugs
Damage dealing was added to previosuly introduced fighting system.
Now after a successful attack entity hp is reduced and if it reaches 0,
entity turns into corpse. If player dies then the game ends. Defense
parameter is a chance of avoiding damage in the attack and power is
how much hit points are lost after successful attack.
(differs form the tutorial) Render order was added to correctly render
entities as sometimes they were covered by corpse.

A bit of code cleaning, mainly for input handler, was performed. But
also for actions to change Entity for Actor in entity types.

FIXED BUGS:
- There was a bug that the (double) death message was printed. It stated
   'Bat is dead' and then 'Agent attacks Remains of Bat'
   which now is fixed.
- Corpse were stacking on top od each other and now the move to
  the adjacent tiles if there are not blocked. Otherwise, they are
  moved outside the map