Add Item class and Health Potions objects
Changes include adding new type of entities - Items. It differs from
enemies by components included. Item has consumable component
instead of ai.

An instance of an Item is a Health Potion that increases agents hp by 4.
It would be possible to use it when health is at maximum. For now items
are visible in game but the interaction between agent and the item is not
possible.

Moreover, health potions are randomly placed on the map. Their number is
specifed in main.

Based on: http://rogueliketutorials.com/tutorials/tcod/v2/part-8